Implemented a Graph based scripting system.
This implements a Graph-based event system. The basic idea is that
simple behaviors, like playing a sound, creating an entity, or adjusting
point values are all implemnented as nodes in a graph. When things
happen in the game, they fire off notifications that nodes need to
update. Each node does its work then sends outputs to any other
nodes that are listening to it which in turn do some work. Any nodes
that have had no change in their inputs don't bother re-running.

Graphs may have multiple instances for if you want to duplicate
behaviors across multiple entities or actors. Each unique instance is
called a GraphState.

Tested on Linux and Android.

Change-Id: I0bed12aca3fe6e0aea936dad8fe4c9694c9fa7e2