Streamlined type registration.
Previously there was a vector of types on the EventSystem structure that
would keep track of the types that have been registered. Turns out this
vector was unnecessary, so it was removed. Type registration itself has
been moved to the TypeRegistry struct, which now stores types statically
rather than using a vector.

Tested on Linux.

Change-Id: I35f7580ba88f975c6bc9cbff391db7ba92676a6a