Void types can now be used as node edges.
It is sometimes useful to be able to mark a node as dirty without
sending it any data (for example, force a PlaySound node to re-evaluate
and play a new sound without altering the location, sound handle, or
gain). The TypeRegistry has been updated to allow for void edges which
carry no data but still force a timestamp to be created to track when
they've been set.

Also did a minor refactor to clean up error messages in io.h while
making it work with void outputs.

Tested on Linux.

Change-Id: I24fcf3d96fcf707cf3bf7dd5d93a1a934eb38021