Added a GraphLoader class.
GraphLoader is a utility class that can help load graphs from data. The
user of the class must override the virtual function 'ParseData' to
interpret the input source. The GraphFactory maintains a map of loaded
files though, and will not bother re-loading a file that was previously
opened.

Tested on Linux.

Change-Id: I854b8d1a574d57ab3b5540b0d5c34666f3b201ab