Fixed node events; renamed inputs and outputs.
Node arguments were broken in that per-graphstate data was being stored
on shared graphs. This change makes it so that node events are sent to
the correct node states, and changes how broadcasters are set up
slightly.

Inputs and outputs have been changed to be NodeArguments, which combines
the roles of the inputs and output classes, and allows access to other
bits of data like Listeners without needing to change function
signatures moving forward.

Tested on Linux.

Change-Id: I4f5dc8697d80c51539aff7e067fc3c2cd629e828