Changed how timestamp initialization works.
Previously the timestamp was initialized to 1 to avoid all nodes
executing the first time Execute was called. (Nodes are considered dirty
when their timestamp is equal to the graph's timestamp, and the nodes
start with a timestamp of 0).

This caused a problem when the nodes set data during initialization,
because that would mark those nodes as dirty.

The fix is to set the timestamp to 0 for initialization, then increment
it to 1 when initialization is complete.

Tested on Linux.

Change-Id: I0584864de811bd782ec962f131be9c50c8f0cb07