Added an IsInputDirty function.
Using IsInputDirty the user can determine which inputs were set since
the last call to Execute. This is useful in a number of situations. For
example, if you have a node to play audio, you might only want to play
when a specific edge is triggered rather than any time any parameter
changes.

Tested on Linux and android.

Change-Id: Id998cfdf364b3171e59b797776db5aea08aeca95