Moved common modules from Zooshi to Breadboard.
Imported a number of common modules from Zooshi into Breadboard in a
separate module_library. Modules that have no dependencies like the math
and string modules are always compiled if the module_library is
compiled. A handful of other modules based on other FPL technologies can
be conditionally included via options in CMakeLists.txt.

Moved a bunch of source files from src/ to src/breadboard. There will
soon be tests and samples in the src/ directory too.

Tested on Linux and Android.
Bug: 24194494
Change-Id: Id36258f6d0e81abbc7664171f936bae0dc02d2c3