Moved common modules into breadboard.
Modules that have no dependencies have been moved into breadboard
itself so that users don't have to include a second library just to get
basic functionality.

Also did some refactoring. EventSystem is now called the ModuleRegistry
because that's what it actually handles, and no longer has anything to
do directly with events. The AddModule function was also renamed
RegisterModule.

Added some using directives to make it so that we don't have to prefix
everything with breadboard::

Tested on Linux.

Change-Id: I5188082e416e93043f7a9ee1ac074fe49eb7c535