Added Doxygen documentation to Breadboard.
Added documentation to all public headers and added a few pages in the
docs folder to show typical usage of the library.

Also added a README and CONTRIBUTING file.

A few files were intentionally skipped a few structures and wrapped them in
@cond/@endcond blocks. These strctures need to be exposed so they can be used
by other header files, but are intended to be for internal use only. They've
been anontated with BREADBOARD_INTERNAL.

Tested on Linux

Change-Id: I8638d4e519bb01d01c94dd6ff652ca3d6715c025