Fixed up build for circular dependencies between CORGI and Breadboard.
The Breadboard module_library build was failing when built as part of
CORGI in the distribution source layout on Linux.  This was due
to CORGI defining Breadboard's location under dependencies/breadboard so
this change overrides this behavior by setting the Breadboard location in
the project.

In addition, this change "fixes" (I'm not sure whether the solution will
work on Windows) recursive builds of Breadboard in the distribution
source layout on Android.

Tested:
Verified CORGI builds all targets (including Breadboard) on Linux and
for Android in the distribution source layout.

Bug: 22510457
Change-Id: Id2f6434a4673227c5b5263e0946c4755f3252db9