Simplified Graphs and Fixed a few errors.
Removed the *DataRef (e.g. TransformDataRef) edge types. They didn't
actually save us any memory or speed things up at all, and they only
made the graphs more complicated. Things that used to take a *DataRef
now just take entities instead.

A lot of graphs had redundant nodes or nodes responding to events they
shouldn't, and this would cause errors like incrementing the target
score sometimes when it shouldn't. These have been cleaned up.

Also added a few nodes. A Velocity node for getting the physics velocity
of an object, a vector length node (useful for getting the magnitude of
the velocity), and an 'if-gate' node that only evaluates the boolean
input when its triggered.

Also fixed an assert error that would occur in the case where a void
node was the last node in the graph. This would result in the graph code
attempting to get the pointer to the non-existant object (which is
normally fine since the pointer would not end up getting dereferenced).
The issue was that the address was at the end of the memory buffer,
which would cause the buffer to assert since it doesn't like it when
things try to access data at offsets >= buffer.size().

Tested on Linux and Android.
Bug: 25279644
Change-Id: Ie94c01da63a46d42348d9f10c58c3a108c100b2d