Added the ability to name node parameter indexes.
Input and output parameters can now be given indexes when defined, which
can also be used when querying the value of the argument. You can also
supply a name and comment for arguments. This currently does nothing,
but can be used in graphical editors and error messages in the future.

Tested on Linux.
Bug: 24300828

Change-Id: I324fcdfc7e5582d37b0e5ce3e35d16c3d5467319